requirements-tests.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --extra=tests --output-file=requirements-tests.txt pyproject.toml
#
attrs==21.4.0
    # via
    #   automat
    #   gather (pyproject.toml)
    #   twisted
    #   virtue
automat==20.2.0
    # via twisted
click==8.1.3
    # via virtue
colorama==0.4.5
    # via virtue
constantly==15.1.0
    # via twisted
coverage[toml]==6.4.1
    # via gather (pyproject.toml)
hyperlink==21.0.0
    # via twisted
idna==3.3
    # via hyperlink
incremental==21.3.0
    # via
    #   gather (pyproject.toml)
    #   twisted
pyhamcrest==2.0.3
    # via gather (pyproject.toml)
pyrsistent==0.18.1
    # via virtue
six==1.16.0
    # via automat
tomli==2.0.1
    # via coverage
twisted==22.4.0
    # via virtue
typing-extensions==4.3.0
    # via twisted
venusian==3.0.0
    # via gather (pyproject.toml)
virtue==2.1.1
    # via gather (pyproject.toml)
zope-interface==5.4.0
    # via twisted

# The following packages are considered to be unsafe in a requirements file:
# setuptools
